refactor(core): extract functions from useEventsStore hook
